move client specific logic into its own file
